added functional for getting info about new users in system added answer functional added control type functional created repositories, services updated permission entity updated folder functional (remove parent folder, added root marker) added new logic to mapping functional updated dtos, response (mapping), request (UserUpdateRequest) added init data and scheme

+++ src/main/java/com/datapath/checklistukraineapp/dao/repository/PermissionRepository.java
@@ -1,11 +1,11 @@
 package com.datapath.checklistukraineapp.dao.repository;
 
-import com.datapath.checklistukraineapp.dao.entity.PermissionEntity;
+import com.datapath.checklistukraineapp.dao.entity.classifier.PermissionEntity;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
 import java.util.Optional;
 
-public interface PermissionRepository extends Neo4jRepository<PermissionEntity, Long> {
+public interface PermissionRepository extends Neo4jRepository<PermissionEntity, Integer> {
 
     Optional<PermissionEntity> findByRole(String role);
 }
